Handle constant integer negation and INT_MIN
Negating a constant now produces an AST integer node with negative
value. AST and tokens are modified to account for INT_MIN appearing in
code, as this value will at least temporarily appear is INT_MAX + 1, so
i32 is not valid data type internally.

+++ src/token/mod.rs
@@ -96,7 +96,16 @@ pub enum TokenSubType {
   Text(Rc<String>),
   FloatNumber(f32),
   DoubleNumber(f64),
-  IntegerNumber(i32),
+  /*
+   *  At token stage, we only deal with positive values; later stages will negate numbers, as they
+   *  have more context on if minus means negate or is part of arithmetic expression
+   *  (for example, a = -45 vs a= 2-45).
+   *
+   *  This means that we must be able to store INT_MAX + 1, when token is actually INT_MIN.
+   *  To do this, use u32 at this stage, and convert to i32 later on when number is known
+   *
+   */
+  IntegerNumber(u64),
   Identifier(Rc<String>),
   BooleanValue(bool),
   FloatType,
